first try at an "isolated blob" diagnostic
A "blob" is obtained by merging bounding boxes that intersect until no
remaining boxes intersect each other.
A padding is added so that boxes that are close together are merged.

To avoid quadratic complexity when many objects are present, the
"intervaltree" library was added, together with the "orderedcontainer"
library on which it depends.

This is not yet entirely satisfactory.